CY-1181: adding depends_on_lifecycle_operation
- Adding also DSL support
- Separating relationships unit tests to multiple files
